Add files via upload
An Interactive and dynamic HR dashboard was developed using Tableau. The dataset was sourced via ChatGPT and meticulously cleaned and organized for accuracy and reliability. Various calculated fields, parameters, and functions were utilized to effectively manipulate the data. Comprehensive visualizations, including heat maps, barbell charts, and bar graphs, provide a clear and insightful overview of HR metrics. The dashboard features user-friendly navigation, connecting to another dashboard via navigation icons and tools, and includes interactive filters to allow users to visualize and obtain the data they desire with ease. The focus is on detailed presentation, ensuring that the information visualized is both concise and accurate.